LayoutTests/pointer-lock/lock-already-locked-expected.txt

Test calling lock when already in a locked state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.webkitPointer.isLocked is false
PASS Lock target1 succeeded.
PASS navigator.webkitPointer.isLocked is true
PASS Lock target1 succeeded again.
PASS navigator.webkitPointer.isLocked is true
PASS expectTarget1Unlock is true
PASS Lock target2 succeeded.
PASS navigator.webkitPointer.isLocked is true
PASS successfullyParsed is true

TEST COMPLETE
